+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/HUD/BackpackEditorHUDV2/AvatarSlot/AvatarSlotComponentModel.cs
@@ -1,0 +1,14 @@
+using System;
+
+namespace DCL.Backpack
+{
+    [Serializable]
+    public class AvatarSlotComponentModel : BaseComponentModel
+    {
+        public string rarity;
+        public string imageUri;
+        public string category;
+        public bool isHidden;
+        public string hiddenBy;
+    }
+}
