+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/HUD/BackpackEditorHUDV2/AvatarSlot/IAvatarSlotComponentView.cs
@@ -1,0 +1,19 @@
+using System;
+
+namespace DCL.Backpack
+{
+    public interface IAvatarSlotComponentView
+    {
+        event Action<string> OnSelectAvatarSlot;
+
+        void SetIsHidden(bool isHidden, string hiddenBy);
+
+        void SetCategory(string category);
+
+        void SetNftImage(string imageUri);
+
+        void SetRarity(string rarity);
+
+        void OnPointerClickOnDifferentSlot();
+    }
+}
