+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/HUD/BackpackEditorHUDV2/BackpackEditorHUDV2.asmdef
@@ -60,7 +60,8 @@
         "GUID:c8ea72e806cd2ab47b539b37895b86b7",
         "GUID:8aa60ff897c74401cb96cdbc583fed0d",
         "GUID:4d3366a36e77f41cfb7436340334e236",
-        "GUID:bab23c9e6b4c4da469d4688d16c1ee9b"
+        "GUID:bab23c9e6b4c4da469d4688d16c1ee9b",
+        "GUID:d076a2c146613441bb926f5b49a88a9f"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
