+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/HUD/BackpackEditorHUDV2/Tests/BackpackEditorHUDV2Tests.asmdef
@@ -32,7 +32,8 @@
         "GUID:3b80b0b562b1cbc489513f09fc1b8f69",
         "GUID:68069f49d86442cd9618861b4d74b1aa",
         "GUID:f51ebe6a0ceec4240a699833d6309b23",
-        "GUID:0c692b52b8dc4a8e9802ddcb8e10f231"
+        "GUID:0c692b52b8dc4a8e9802ddcb8e10f231",
+        "GUID:d076a2c146613441bb926f5b49a88a9f"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
