--- unity-renderer/Assets/UIComponents/Scripts/Components/AvatarSlot/AvatarSlotComponentModel.cs
@@ -1,11 +1,0 @@
-using System;
-
-[Serializable]
-public class AvatarSlotComponentModel : BaseComponentModel
-{
-    public string rarity;
-    public string imageUri;
-    public string category;
-    public bool isHidden;
-    public string hiddenBy;
-}
